+++ client/rpc/src/main/kotlin/net/corda/client/rpc/internal/RPCClient.kt
@@ -1,7 +1,7 @@
 package net.corda.client.rpc.internal
 
 import net.corda.core.crypto.random63BitValue
-import net.corda.core.logElapsedTime
+import net.corda.core.internal.logElapsedTime
 import net.corda.core.messaging.RPCOps
 import net.corda.core.serialization.SerializationContext
 import net.corda.core.serialization.SerializationDefaults