+++ node/src/integration-test/kotlin/net/corda/node/services/DistributedServiceTests.kt
@@ -1,6 +1,6 @@
 package net.corda.node.services
 
-import net.corda.core.bufferUntilSubscribed
+import net.corda.core.internal.bufferUntilSubscribed
 import net.corda.core.contracts.Amount
 import net.corda.core.contracts.POUNDS
 import net.corda.core.identity.Party