+++ node/src/main/kotlin/net/corda/node/services/statemachine/StateMachineManager.kt
@@ -8,8 +8,8 @@ import com.esotericsoftware.kryo.KryoException
 import com.google.common.collect.HashMultimap
 import com.google.common.util.concurrent.MoreExecutors
 import net.corda.core.internal.ThreadBox
-import net.corda.core.bufferUntilSubscribed
 import net.corda.core.concurrent.CordaFuture
+import net.corda.core.internal.bufferUntilSubscribed
 import net.corda.core.crypto.SecureHash
 import net.corda.core.crypto.random63BitValue
 import net.corda.core.flows.FlowException