+++ node/src/main/kotlin/net/corda/node/shell/InteractiveShell.kt
@@ -7,7 +7,6 @@ import com.fasterxml.jackson.databind.*
 import com.fasterxml.jackson.databind.module.SimpleModule
 import com.fasterxml.jackson.dataformat.yaml.YAMLFactory
 import com.google.common.io.Closeables
-import net.corda.core.*
 import net.corda.core.concurrent.CordaFuture
 import net.corda.core.flows.FlowInitiator
 import net.corda.core.flows.FlowLogic
@@ -17,6 +16,7 @@ import net.corda.core.internal.concurrent.openFuture
 import net.corda.core.internal.createDirectories
 import net.corda.core.internal.div
 import net.corda.core.internal.write
+import net.corda.core.internal.*
 import net.corda.core.messaging.CordaRPCOps
 import net.corda.core.messaging.StateMachineUpdate
 import net.corda.core.internal.Emoji